region_cache: move store related code to a standalone file
